Refactored some code; organized editDomainlist in preparation for use (issue #175)
